Refactor and implement process in Pipeline
