+++ main.py
@@ -9,12 +9,13 @@ def main():
     parser.add_argument("video_capture_address")
     args = parser.parse_args()
     cap = cv2.VideoCapture(args.video_capture_address)  # type: ignore
-    pipeline = Pipeline(args)
+    pipeline = Pipeline()
     if not cap.isOpened():
         print("Can't open camera")
         exit()
     while True:
         ret, image = cap.read()
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGBA)  # type: ignore
         if not ret:
             print("Can't receive frame (stream end?). Exiting ...")
             break
